sync

+++ app/src/main/java/com/whale/shapex2d/interfaces/Movable.java
@@ -12,8 +12,8 @@ public interface Movable {
     Vec2D getPosition();
     Vec2D getVelocity();
     void setVelocity(Vec2D velocity);
-    void setRadius(int r);
-    int getRadius();
+    void setRadius(double r);
+    double getRadius();
     int move();
     int move(int velX, int velY);
     boolean intersects(int x, int y);
